refactor: web와 console용 패키지 재분리
