+++ src/main/java/chess/web/dao/MemberDao.java
@@ -1,9 +1,9 @@
-package chess.dao;
+package chess.web.dao;
 
-import chess.domain.Member;
+import chess.web.Member;
 import java.util.List;
 
-public interface MemberDaoInterface {
+public interface MemberDao {
     void removeById(String id);
 
     void save(Member member);
