+++ src/main/java/chess/web/service/ChessService.java
@@ -1,4 +1,4 @@
-package chess.service;
+package chess.web.service;
 
 public class ChessService {
 }
